requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
colorama==0.4.4
cryptography==3.4.7
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.5
entrypoints==0.3
firebase-admin==5.0.1
gcloud==0.18.3
git-clone==1.0.6
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.2
googleapis-common-protos==1.53.0
grpcio==1.39.0
httplib2==0.19.1
idna==3.2
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jwcrypto==0.9.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
oauth2client==4.1.3
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.10.0
pyparsing==2.4.7
Pyrebase4==4.5.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-firebase==1.2
python-jwt==3.3.0
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.1
terminado==0.11.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
uuid==1.30
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
